+++ src/app.js
@@ -27,6 +27,7 @@ const defaultConfig = {
     '--color=always',
     '--pretty=format:%C(yellow)%h %Creset%s%Cblue [%cn] %Cred%d ',
   ],
+  sort: '-committerdate',
   themeColor: '#FFA66D',
 };
 
@@ -199,7 +200,7 @@ export const start = args => {
    * @param  remoteList {Array} Unique remotes for current project
    * @return {String}
    */
-  function getPrevRemote(currentRemote, remoteList) {
+  const getPrevRemote = (currentRemote, remoteList) => {
     let currIndex = remoteList.indexOf(currentRemote);
 
     if (currIndex > 0) {
@@ -211,7 +212,7 @@ export const start = args => {
     refreshTable(currentRemote);
 
     return currentRemote;
-  }
+  };
 
   /**
    * Cycle to next remote
@@ -220,7 +221,7 @@ export const start = args => {
    * @param  remoteList {Array} Unique remotes for current project
    * @return {String}
    */
-  function getNextRemote(currentRemote, remoteList) {
+  const getNextRemote = (currentRemote, remoteList) => {
     let currIndex = remoteList.indexOf(currentRemote);
 
     if (currIndex < remoteList.length - 1) {
@@ -232,14 +233,14 @@ export const start = args => {
     refreshTable(currentRemote);
 
     return currentRemote;
-  }
+  };
 
   /**
    * Update current screen with current remote
    *
    * @param {String} currentRemote Current displayed remote
    */
-  function refreshTable(currentRemote = 'heads') {
+  const refreshTable = (currentRemote = 'heads') => {
     branchTable.setData([
       ['', 'Remote', 'Ref Name'],
       ...branchPayload[currentRemote],
@@ -249,5 +250,5 @@ export const start = args => {
 
     loading.stop();
     screen.render();
-  }
+  };
 };