+++ CHANGELOG.md
@@ -24,11 +24,24 @@ Each issue fixed should contain one bullet summarizing the work done.
 ### Security
 ```
 
-
-
 <hr />
 
 ## [Unreleased]
+
+
+
+## [[0.9.2]](https://github.com/jwu910/check-it-out/releases/tag/v0.9.2) - 2018-07-19
+### Notes
+Some minor changes in preparation for 1.0.0
+
+### Added
+* [CIO-203](https://github.com/jwu910/check-it-out/issues/203) Add configuration for sort date option
+
+### Changed
+* [CIO-198](https://github.com/jwu910/check-it-out/issues/198) Change function declarations to const
+
+### Fixed
+* [CIO-120](https://github.com/jwu910/check-it-out/issues/120) Fix table width to show scrollbar
 
 
 ## [[0.9.1]](https://github.com/jwu910/check-it-out/releases/tag/v0.9.1) - 2018-07-16